+++ packages/browser-wallet/src/popup/store/settings.ts
@@ -25,6 +25,7 @@ export const encryptedSeedPhraseAtom = atomWithChromeStorage<EncryptedData | und
 );
 export const themeAtom = atomWithChromeStorage<Theme>(ChromeStorageKey.Theme, Theme.Light);
 export const uiStyleAtom = atomWithChromeStorage<UiStyle>(ChromeStorageKey.UiStyle, UiStyle.Old, true);
+export const memoryStoreUpdateAtom = atomWithChromeStorage<number>(ChromeStorageKey.MemoryStoreUpdate, 0, true);
 export const hasBeenOnBoardedAtom = atomWithChromeStorage<boolean>(ChromeStorageKey.HasBeenOnboarded, false, true);
 export const hasBeenSavedSeedAtom = atomWithChromeStorage<boolean>(ChromeStorageKey.HasBeenSavedSeed, true, false);
 
@@ -85,12 +86,6 @@ export const grpcClientAtom = atom<ConcordiumGRPCClient>((get) => {
 
 export const acceptedTermsAtom = atomWithChromeStorage<AcceptedTermsState | undefined>(
     ChromeStorageKey.AcceptedTerms,
-    undefined,
-    true
-);
-
-export const sessionPasscodeAtom = atomWithChromeStorage<string | undefined>(
-    ChromeStorageKey.Passcode,
     undefined,
     true
 );